+++ controllers/contacts.js
@@ -1,15 +1,5 @@
-const Joi = require("joi");
-
 const contacts = require("../models/contacts");
 const { HttpError, ctrlWrapper } = require("../helpers");
-
-const newSchema = Joi.object({
-  name: Joi.string().min(3).max(30).required(),
-  email: Joi.string()
-    .email({ tlds: { deny: ["ru", "by", "hu"] } })
-    .required(),
-  phone: Joi.string().min(6).max(20).required(),
-});
 
 const getAllContacts = async (req, res) => {
   const result = await contacts.listContacts();
@@ -25,10 +15,6 @@ const getContactById = async (req, res) => {
 };
 
 const createContact = async (req, res) => {
-  const { error } = newSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
   const result = await contacts.addContact(req.body);
   res.status(201).json(result);
 };
@@ -42,10 +28,6 @@ const deleteContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
-  const { error } = newSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
   const result = await contacts.updateContact(req.params.contactId, req.body);
   if (!result) {
     throw HttpError(404, "Contact with this id not found");
