+++ routes/api/contacts.js
@@ -1,6 +1,8 @@
 const express = require("express");
 
 const ctrl = require("../../controllers/contacts");
+const schemas = require("../../schemas/contacts");
+const { validateBody } = require("../../middlewares");
 
 const router = express.Router();
 
@@ -8,10 +10,10 @@ router.get("/", ctrl.getAllContacts);
 
 router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", ctrl.createContact);
+router.post("/", validateBody(schemas.newSchema), ctrl.createContact);
 
 router.delete("/:contactId", ctrl.deleteContact);
 
-router.put("/:contactId", ctrl.updateContact);
+router.put("/:contactId", validateBody(schemas.newSchema), ctrl.updateContact);
 
 module.exports = router;
